Retail Platform

+++ src/gold/build_marts.py
@@ -1,18 +1,46 @@
-from pyspark.sql import SparkSession
+from src.common.spark_session import get_spark
+from pyspark.sql.functions import col, sum as spark_sum
 import glob
 import shutil
 import os
 
-spark = SparkSession.builder \
-    .appName("Gold Layer") \
-    .master("local[*]") \
-    .getOrCreate()
+spark = get_spark()
 
-silver_df = spark.read.option("header", True) \
-    .csv("silver/customer_sales")
+# Read Data
+customers = spark.read.parquet(
+    "silver/customer_sales"
+)
 
-gold_df = silver_df.groupBy("category") \
-    .sum("sales_amount")
+orders = spark.read.parquet(
+    "bronze/orders"
+)
+
+products = spark.read.parquet(
+    "bronze/products"
+)
+
+# Join Data
+sales = orders.join(
+    customers,
+    "customer_id"
+).join(
+    products,
+    "product_id"
+)
+
+# Calculate Revenue
+sales = sales.withColumn(
+    "revenue",
+    col("quantity").cast("int") *
+    col("price").cast("int")
+)
+
+# Aggregate for Power BI
+gold_df = sales.groupBy(
+    "category"
+).agg(
+    spark_sum("revenue").alias("total_sales")
+)
 
 gold_output = "gold/customer_sales"
 
@@ -22,16 +50,23 @@ gold_df.coalesce(1) \
     .option("header", True) \
     .csv(gold_output)
 
-csv_file = glob.glob(
+# Create single CSV file
+csv_files = glob.glob(
     f"{gold_output}/part-*.csv"
-)[0]
-
-if os.path.exists("customer_sales.csv"):
-    os.remove("customer_sales.csv")
-
-shutil.copy(
-    csv_file,
-    "customer_sales.csv"
 )
 
+if len(csv_files) > 0:
+
+    if os.path.exists("customer_sales.csv"):
+        os.remove("customer_sales.csv")
+
+    shutil.copy(
+        csv_files[0],
+        "customer_sales.csv"
+    )
+
 print("Gold Layer Completed")
+
+gold_df.show()
+
+spark.stop()
